fix: cloudnet luckperms does not load on bungeecord & nukkit (#1547)
### Motivation
Due to the usage of platform inject we need to specify the name of the
resulting plugin.yml. Bungeecord and nukkit lack that setting and
therefore try to use the bukkit plugin.yml.

### Modification
Added the pluginFileName for nukkit & bungeecord according to their
plugin loaders.

### Result
The cloudnet luckperms plugin works properly on these platforms.