fix: exception when interacting with npc inventory (#1554)
### Motivation
An NPCs inventory is only created when the NPC is spawned. However,
we only spawn NPCs if the chunk they are located in is loaded. As a
result, there are NPCs that are tracked but do not yet have an inventory.
When interacting with an already spawned NPC, a check is performed to
determine if the clicked inventory belongs to an NPC. If an NPC is not yet
spawned and therefore its inventory has not yet been created, checking
whether the inventory belongs to this NPC will result in a
NullPointerException.

### Modification
Mark the method that returns the NPC's inventory as nullable and
make the inventory comparison null aware.

### Result
Clicking on NPC selector items no longer leads to exceptions due to
NPCs without inventory.