fix: service copy without default template (#1558)
### Motivation
When copying a service without providing a template we try to fallback
to the "<TaskName>/default" template. But our fallback logic is not
working correctly and causing an NPE.

### Modification
We now only accept a single service when copying. If no default template
exists and no template is provided we respond with an message indicating
the missing template.

### Result
The `service <Service> copy` Command is working again.

##### Other context
Fixes #1550