fix: command parsing and executing caused console dead lock (#1565)
### Motivation
Executing a lot of commands while e.g. services are starting (printing
messages to the console) can lead to dead locks. This issue was probably
introduced with the cloud 2.0 migration. Prior to cloud 2.0 the
framework used to run suggestions on the caller thread. With cloud 2.0 a
shared thread pool for suggestion, parsing and executing can be used.
This then leads to dead locks.

### Modification
Explicitly set the thread pool for parsing & executing only. Suggestions
are processed on the caller thread (thus restoring cloud 1.0 behavior.

### Result
No dead-locks of the console.