fix: use AsyncPlayerConfgiurationEvent for permission checks on minestom
### Motivation
Maintainance kicks on minestom kick the player after he has joined. This
causes join/leave messages for that player to be visible, even though he
was disallowed to join.

### Modification
I changed the event with the permission check from PlayerSpawnEvent to
AsyncPlayerConfigurationEvent.
Permission systems have time to load the permission data before that
time by using `AsyncPlayerPreLoginEvent`, and the player won't be in the
world yet. Perfect time for permission checks to deny login.

### Result
The player properly gets denied on "login", not kicked right after he
was able to join.